requirements.txt
fastapi
uvicorn
torch
torchvision
faiss-cpu
Pillow
woocommerce
pymysql
    